ADDED, starting to add tests for database
Added two test for testing the PortfolioDAO, maybe I should test the repository to test the two clases at the same time

+++ app/src/main/java/com/github/ricardopolit/sipmanager/data/portfolio/Portfolio.kt
@@ -24,7 +24,7 @@ data class Portfolio(
         var dateFinish: Long,
 
         @ColumnInfo(name = "total_deposits")
-        var totalDeposits: Int = 0,
+        var totalDeposits: Float = 0.0F,
 
         @ColumnInfo(name = "earnings")
         val earnings: Float = 0.0F,
